+++ src/main/java/leets/weeth/global/sas/config/authentication/ProviderAwareEntryPoint.java
@@ -1,9 +1,12 @@
-package leets.weeth.global.config;
+package leets.weeth.global.sas.config.authentication;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import leets.weeth.global.auth.authentication.ErrorMessage;
+import leets.weeth.global.common.response.CommonResponse;
+import leets.weeth.global.sas.application.property.OauthProviderProperties;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
@@ -12,26 +15,27 @@ import org.springframework.web.util.UriComponentsBuilder;
 import java.io.IOException;
 import java.util.Optional;
 
-@Slf4j
 @Component
 @RequiredArgsConstructor
 public class ProviderAwareEntryPoint implements AuthenticationEntryPoint {
 
     private final OauthProviderProperties oauthProps;
 
+    private static final String PROVIDER_PARAMETER = "provider";
+    private static final String KAKAO = "kakao";
+
     @Override
     public void commence(HttpServletRequest req, HttpServletResponse res, AuthenticationException ex) throws IOException {
 
         // 1) provider 파라미터 파싱 (없으면 kakao 기본)
-        String provider = Optional.ofNullable(req.getParameter("provider"))
-                .orElse("kakao");
+        String provider = Optional.ofNullable(req.getParameter(PROVIDER_PARAMETER))
+                .orElse(KAKAO);
 
         OauthProviderProperties.Provider cfg = oauthProps.getProviders().get(provider);
 
         // 2) 잘못된 provider 처리
         if (cfg == null) {
-            log.warn("Unknown provider: {}", provider);
-            res.sendError(HttpServletResponse.SC_BAD_REQUEST, "unknown provider");
+            setResponse(res);
             return;
         }
 
@@ -45,5 +49,14 @@ public class ProviderAwareEntryPoint implements AuthenticationEntryPoint {
                 .toUriString();
 
         res.sendRedirect(redirect);
+    }
+
+    private void setResponse(HttpServletResponse response) throws IOException {
+        response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+        response.setContentType("application/json");
+        response.setCharacterEncoding("UTF-8");
+
+        String message = new ObjectMapper().writeValueAsString(CommonResponse.createFailure(ErrorMessage.SC_BAD_REQUEST_PROVIDER.getCode(), ErrorMessage.SC_BAD_REQUEST_PROVIDER.getMessage()));
+        response.getWriter().write(message);
     }
 }
